# Practica3/src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Coforge Aplication
  description: |
    Coforge Aplication
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
  /api/employees:
    post:
      tags:
        - Employees
      description: |
        Add an employee to the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeDTO'
      responses:
        '204':
          description: Employee added successfully
        '400':
          description: Error in the request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Error inserting employee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error
    get:
      tags:
        - Employees
      description: |
        Get all employees in the system.
      responses:
        '200':
          description: Employees retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeDTO'
        '500':
          description: Internal server error
  /api/employees/{code}:
    get:
      tags:
        - Employees
      description: |
        Get information about a specified employee by employee code.
      parameters:
        - name: code
          in: path
          description: Employee code to be queried
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Employee found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDTO'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error
    put:
      tags:
        - Employees
      description: |
        Update data of an employee in the system.
      parameters:
        - name: code
          in: path
          description: Employee code to be updated
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UEmployeeDTO'
      responses:
        '204':
          description: Employee updated successfully
        '400':
          description: Values for employee are wrong
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Error updating employee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error
    delete:
      tags:
        - Employees
      description: |
        Delete an employee from the system by employee code.
      parameters:
        - name: code
          in: path
          description: Employee code to be deleted
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Employee deleted successfully
        '409':
          description: Error deleting employee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error
...
  /api/assignments:
    post:
      tags:
        - Assignments
      description: |
        Add a new assignment in the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentDTO'
      responses:
        default:
          description: Unexpected error
        '204':
          description: Assignment inserted successfully
        '400':
          description: Values for assignment are wrong
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Error inserting the assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Assignments
      description: |
        Get all assignments in the system.
      responses:
        default:
          description: Unexpected error
        '200':
          description: Assignments retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Assignment'
        '500':
          description: Internal server error
  /api/assignments/{code}:
    get:
      tags:
        - Assignments
      description: |
        Get information about an assignment of a specific employee by their code.
      parameters:
        - name: code
          in: path
          description: Employee code whose assignment is to be displayed
          required: true
          schema:
            type: integer
      responses:
        default:
          description: Unexpected error
        '200':
          description: Assignment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '404':
          description: Assignment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Assignments
      description: |
        Update an assignment by employee code.
      parameters:
        - name: code
          in: path
          description: Employee code whose assignment is to be updated
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentDTO'
      responses:
        default:
          description: Unexpected error
        '204':
          description: Assignment updated successfully
        '400':
          description: Values for assignment are wrong
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Assignment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Error updating assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Assignments
      description: |
        Delete an existing assignment from the system by employee code.
      parameters:
        - name: code
          in: path
          description: Employee code whose assignment is to be deleted
          required: true
          schema:
            type: integer
      responses:
        default:
          description: Unexpected error
        '204':
          description: Assignment deleted successfully
        '409':
          description: Error deleting assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    AssignmentDTO:
      type: object
      properties:
        employeeCode:
          type: integer
        supervisorCode:
          type: integer
        assignmentInfo:
          properties:
            remark:
              type: string
            percentage:
              type: string
            startDate:
              type: string
            endDate:
              type: string
        project:
          properties:
            oldCode:
              type: string
            newCode:
              type: string
        practiceName:
          type: string
      required:
        - name
    EmployeeDTO:
      type: object
      properties:
        code:
          type: integer
        name:
          type: string
        role:
          type: string
        practice:
          type: string
      required:
        - code
        - name
        - role
    UEmployeeDTO:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        practice:
          type: string
...
    Assignment:
      type: object
      properties:
        employee:
          $ref: '#/components/schemas/EmployeeDTO'
        supervisor:
          properties:
            code:
              type: integer
            name:
              type: string
        practiceName:
          type: string
        project:
          properties:
            oldCode:
              type: string
            newCode:
              type: string
        assignmentInfo:
          properties:
            code:
              type: integer
            remark:
              type: string
            percentage:
              type: string
            startDate:
              type: string
            endDate:
              type: string
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        reason:
          type: string